feat: room delete

+++ src/main/java/com/trianglechoke/codesparring/room/service/RoomService.java
@@ -1,5 +1,8 @@
 package com.trianglechoke.codesparring.room.service;
 
+import static com.trianglechoke.codesparring.exception.ErrorCode.ROOM_NOT_FOUND;
+
+import com.trianglechoke.codesparring.exception.MyException;
 import com.trianglechoke.codesparring.quiz.entity.Quiz;
 import com.trianglechoke.codesparring.room.dao.RoomRepository;
 import com.trianglechoke.codesparring.room.dto.RoomDTO;
@@ -7,6 +10,8 @@ import com.trianglechoke.codesparring.room.entity.Room;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class RoomService {
@@ -26,11 +31,24 @@ public class RoomService {
 
     /* 대기방 삭제 */
     public void removeRoom(Long roomNo) {
-        repository.deleteById(roomNo);
+        Optional<Room> room = repository.findById(roomNo);
+
+        if (room.isPresent()) {
+            repository.deleteById(roomNo);
+        } else {
+            throw new MyException(ROOM_NOT_FOUND);
+        }
     }
 
     /* 대기방 수정 */
-    public void modifyRoom() {}
+    public void modifyRoom(Long roomNo) {
+        Optional<Room> room = repository.findById(roomNo);
+
+        if (room.isPresent()) {
+        } else {
+            throw new MyException(ROOM_NOT_FOUND);
+        }
+    }
 
     /* 대기방 상세 조회 */
     /* 대기방 목록 조회 */
